backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.10.0
beautifulsoup4==4.12.3
requests==2.31.0
PyPDF2==3.0.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
alembic==1.13.1
email-validator==2.3.0
httpx==0.27.2
bcrypt==3.2.2
